+++ release_notes.md
@@ -1,6 +1,12 @@
 PayPal Python SDK release notes
 ============================
 
+v1.6.0
+-----
+* Webhook and Webhook events creation and management supported
+* Parse webhook events and return the appropriate resource
+* Verification that webhook events are unaltered and originate from PayPal
+
 v1.5.0
 -----
 * Payment Experience customizaton feature added via API for Web Profiles